tpfinal/Codigo/api/requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio>=4.0.0
aiohttp>=3.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asttokens==3.0.0
astunparse==1.6.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
comm==0.2.3
fastapi>=0.100.0
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.9.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.75.1
grpcio-tools==1.75.1
h5py==3.14.0
idna==3.10
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kagglehub==0.3.13
keras==3.11.3
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==11.3.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
protobuf==6.32.1
psutil==7.1.0
psycopg>=3.1.0
psycopg-binary>=3.1.0
pure_eval==0.2.3
python-multipart>=0.0.6
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
redis>=5.0.0
requests==2.32.5
rich==14.2.0
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
sentry-sdk==2.41.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio>=1.3.0
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette>=0.35.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn>=0.25.0
wandb==0.22.2
wcwidth==0.2.14
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3